added client class, socket doesn't work

+++ lib/packets.dart
@@ -2,4 +2,4 @@ library packets;
 
 import 'dart:typed_data';
 
-const Uint16 ENTER_QUEUE = 0x00;
+const int ENTER_QUEUE = 0x00;
